mini-app/backend/requirements.txt
# Backend requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiogram>=3.2.0
pyrogram>=2.0.0
tgcrypto
python-dotenv
httpx>=0.25.0
pydantic>=2.0.0
groq>=0.4.0
